Move libopencm3 dependent function into generic interfaces
