+++ power_panic.hpp
@@ -1,0 +1,4 @@
+#pragma once
+
+/// Active wait until power panic signal is disabled
+void WaitForEndOfPowerPanic();
